update: move web3 wallet components to the bal component library
